+++ packages/terra-data-grid/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 1.20.0 - (March 14, 2024)
+
 * Added
   * Added `event` object to `onCellSelect` callback.